requirements.txt
numpy==1.17.2
opencv-python==4.1.1.26
PyQt5==5.13.0
PyQt5-sip==4.19.19
PyQtChart==5.12
pyserial==3.4